requirements.txt
# ── Web framework ──────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── Supabase & HTTP ────────────────────────────────────────────
supabase>=2.4.0
httpx>=0.27.0

# ── Auth & config ──────────────────────────────────────────────
python-dotenv>=1.0.0
pydantic>=2.7.0

# ── ML core ────────────────────────────────────────────────────
scikit-learn>=1.4.0
numpy>=1.26.0
scipy>=1.13.0
pandas>=2.2.0
joblib>=1.4.0


# ── Transformer embeddings (optional but recommended) ──────────
# Enables cold-start predictions and richer feature vectors.
# ~90MB download on first use (model cached in ~/.cache/torch).
# Remove this line if you want a lighter install — the categoriser
# falls back gracefully to TF-IDF-only mode without it.
sentence-transformers>=3.0.0

# ── Gmail fetch (if used) ──────────────────────────────────────
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.128.0


# ── CRON Jobs ──────────────────────────────────────
apscheduler 
pytz

# ── Redis (for CRON job locking) ───────────────────────────────
celery 
redis

